requirements.txt
# Python课程实践项目 - 依赖包列表
# 安装方法: pip install -r requirements.txt

# GUI框架
customtkinter>=5.2.1        # 现代化Tkinter界面
Pillow>=12.0.0              # 图像处理支持

# 数据处理
pandas>=2.2.3               # 数据分析（兼容 Python 3.13）
numpy>=2.1.0                # 数值计算（兼容 Python 3.13）

# 数据可视化
matplotlib>=3.8.2           # 绘图库
seaborn>=0.13.0             # 高级统计图表

# 网络通信
aiohttp>=3.9.1              # 异步HTTP（可选）

# 数据库（如果使用 MySQL/PostgreSQL）
# pymysql>=1.1.0            # MySQL 数据库
# psycopg2>=2.9.9           # PostgreSQL 数据库

# 加密与安全
bcrypt>=4.1.2               # 密码哈希

# 配置文件
pyyaml>=6.0.1                # YAML 配置文件解析

# 工具库
python-dateutil>=2.8.2       # 日期处理
colorlog>=6.8.0              # 彩色日志输出

# Excel 支持（可选）
openpyxl>=3.1.2              # Excel 文件读写

# 开发工具（可选）
pytest>=7.4.3                # 单元测试
black>=23.12.0               # 代码格式化
